feat: compute jshint report and crap score
