+++ package.json
@@ -32,7 +32,11 @@
         "node": ">=18.15.0",
         "npm": ">=9.5.0"
     },
+    "dependencies": {
+        "jshint": "^2.13.6"
+    },
     "devDependencies": {
+        "@types/jshint": "^2.12.1",
         "prettier": "^2.8.4",
         "ts-jest": "^29.0.5",
         "typescript": "^4.9.5"
